requirements.txt
Django==1.11.11
django-filter==1.1.0
django-widget-tweaks==1.4.1
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
